Add back the JSON DATA
The JSON DATA section is used to add the CHANGE_FROM_PREVIOUS_WEEK for
each voice of the report.

The code was removed because the generate JSON would get the issue
content past the allowed limit.

With these changes we are not adding the JSON data if it will bring the
issue chart past the allowed limit.